refactor: extract relay config from nostr-helper-functions into dedicated module
